refactor(recipe): migrate form to client component
